Extract sizing in a separate SizeCalculable protocol
