+++ Sources/Impl/PinLayoutImpl+Size.swift
@@ -112,6 +112,7 @@ extension PinLayout {
     }
     #endif
 
+    @available(OSX 10.10, *)
     public func sizeToFit(_ fitType: FitType) -> PinLayout {
         return setAdjustSizeType(fitType.toAdjustSizeType(), { return "sizeToFit(\(fitType.description))" })
     }
